refactor: add ThemePicker and FormsWrapper (#11)
