+++ apps/web/app/(auth)/theme-picker.tsx
@@ -1,0 +1,41 @@
+'use client';
+
+import { useTheme } from 'next-themes';
+
+import { Button } from '@/components/ui/button';
+import {
+  DropdownMenu,
+  DropdownMenuTrigger,
+  DropdownMenuContent,
+  DropdownMenuItem
+} from '@/components/ui/dropdown-menu';
+
+import { Icons } from '@/components/icons';
+
+import { capitalize } from '@/lib/utils';
+
+export function ThemePicker() {
+  const { setTheme, themes } = useTheme();
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button
+          size="icon"
+          variant="outline"
+          className="absolute right-8 top-8"
+        >
+          <Icons.lightMode className="size-4 rotate-0 scale-100 dark:-rotate-90 dark:scale-0" />
+          <Icons.darkMode className="absolute size-4 rotate-90 scale-0 dark:rotate-0 dark:scale-100" />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent align="end">
+        {themes.map((theme) => (
+          <DropdownMenuItem key={theme} onClick={() => setTheme(theme)}>
+            {capitalize(theme)}
+          </DropdownMenuItem>
+        ))}
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
